+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.29 (January 4th, 2022):
+* [Enhancement] Support game's UI scale in all dialogs.
+* [Enhancement] Reduce the EVA canister dry mass to 1kg.
+* [Fix] Total mass is not ajdusted when fuel is expended from the EVA tank.
+* [Fix] Recalculate the resources mass on load to fix wrong state of the EVA tank.
+* [Fix #394] Refuel tank throws error on refuel action.
+
 # 1.28 (January 15th, 2021):
 * [Fix #378] Dropped parts cannot be moved or trigger physics.
 * [Fix #379] KIS functionality interferes with the stock construction mode.